sql-service/cloud-functions/aqi_to_postgres/requirements.txt
google-cloud-storage>=2.14.0
cloud-sql-python-connector[pg8000]>=1.6.0
SQLAlchemy>=2.0.0
pg8000>=1.29.0
python-dateutil>=2.8.2
pytz>=2023.3
functions-framework>=3.3.0